% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{download_GBD}
\alias{download_GBD}
\title{download GBD data from GBD download website}
\usage{
download_GBD(url, dest)
}
\arguments{
\item{dest}{}
}
\description{
download GBD data from GBD download website
}
